Fix mobile chat chrome controls
