+++ src/app/shell/WindowTitleBar.tsx
@@ -1,9 +1,8 @@
 import { getCurrentWindow } from "@tauri-apps/api/window";
-import { Home, Maximize2, Minus, PanelLeft, PanelLeftClose, Square, X } from "lucide-react";
+import { Maximize2, Minus, Square, X } from "lucide-react";
 import { useCallback, useEffect, useMemo, useState, type MouseEvent as ReactMouseEvent } from "react";
 import { cn } from "../../shared/lib/utils";
-import { useChatStore } from "../../shared/stores/chat.store";
-import { useUIStore } from "../../shared/stores/ui.store";
+import { ChatTitleControls } from "./ChatTitleControls";
 import { PanelNavButtons } from "./PanelNavButtons";
 
 type DesktopPlatform = "darwin" | "windows" | "linux";
@@ -41,10 +40,6 @@ export function WindowTitleBar({
   const platform = useMemo(inferDesktopPlatform, []);
   const [isMaximized, setIsMaximized] = useState(false);
   const appWindow = useMemo(() => (isTauriRuntime() ? getCurrentWindow() : null), []);
-  const setActiveChatId = useChatStore((s) => s.setActiveChatId);
-  const sidebarOpen = useUIStore((s) => s.sidebarOpen);
-  const toggleSidebar = useUIStore((s) => s.toggleSidebar);
-  const closeAllDetails = useUIStore((s) => s.closeAllDetails);
 
   const refreshMaximized = useCallback(() => {
     if (!appWindow) return;
@@ -96,16 +91,6 @@ export function WindowTitleBar({
   const toggleMaximizeFromDragRegion = useCallback(() => {
     runWindowAction("maximize");
   }, [runWindowAction]);
-  const goHome = useCallback(() => {
-    setActiveChatId(null);
-    closeAllDetails();
-    onGoHome?.();
-  }, [closeAllDetails, onGoHome, setActiveChatId]);
-  const openProfessorMari = useCallback(() => {
-    setActiveChatId(null);
-    closeAllDetails();
-    onOpenProfessorMari?.();
-  }, [closeAllDetails, onOpenProfessorMari, setActiveChatId]);
   const controlActions = platform === "darwin" ? (["close", "minimize", "maximize"] as const) : (["minimize", "maximize", "close"] as const);
   const controls = (
     <div
@@ -155,66 +140,12 @@ export function WindowTitleBar({
       onDoubleClick={toggleMaximizeFromDragRegion}
     >
       {platform === "darwin" && controls}
-      <div className="mari-chat-title-controls flex h-full shrink-0 items-center gap-1.5 pl-2.5 pr-0">
-        <button
-          type="button"
-          onClick={toggleSidebar}
-          onMouseDown={(event) => event.stopPropagation()}
-          onDoubleClick={(event) => event.stopPropagation()}
-          data-tour="sidebar-toggle"
-          className={cn(
-            "mari-titlebar-action relative rounded-md p-1.5 transition-all duration-200",
-            sidebarOpen
-              ? "mari-titlebar-action-active text-[color-mix(in_srgb,var(--primary)_54%,var(--muted-foreground))] [&>svg]:stroke-[2.3]"
-              : "text-[var(--muted-foreground)] hover:text-[var(--primary)]",
-          )}
-          title={sidebarOpen ? "Close chats" : "Open chats"}
-          aria-label={sidebarOpen ? "Close chats" : "Open chats"}
-          aria-pressed={sidebarOpen}
-        >
-          {sidebarOpen ? <PanelLeftClose size="0.875rem" /> : <PanelLeft size="0.875rem" />}
-          {sidebarOpen && (
-            <span className="absolute -bottom-0.5 left-1/2 h-0.5 w-3 -translate-x-1/2 rounded-full bg-gradient-to-r from-teal-500 to-cyan-500" />
-          )}
-        </button>
-        <button
-          type="button"
-          onClick={goHome}
-          onMouseDown={(event) => event.stopPropagation()}
-          onDoubleClick={(event) => event.stopPropagation()}
-          className="mari-titlebar-action rounded-md p-1.5 text-[var(--muted-foreground)] transition-all duration-200 hover:text-[var(--primary)]"
-          title="Home"
-          aria-label="Home"
-        >
-          <Home size="0.875rem" />
-        </button>
-        <button
-          type="button"
-          onClick={openProfessorMari}
-          onMouseDown={(event) => event.stopPropagation()}
-          onDoubleClick={(event) => event.stopPropagation()}
-          className={cn(
-            "mari-titlebar-action relative rounded-md p-1 transition-all duration-200",
-            professorMariOpen
-              ? "mari-titlebar-action-active text-[color-mix(in_srgb,var(--primary)_54%,var(--muted-foreground))]"
-              : "text-[var(--muted-foreground)] hover:text-[var(--primary)]",
-          )}
-          title="Professor Mari"
-          aria-label="Professor Mari"
-          aria-pressed={professorMariOpen}
-        >
-          <img
-            src="/sprites/mari/Mari_profile.png"
-            alt=""
-            className="h-[1.125rem] w-[1.125rem] rounded-md object-cover"
-            draggable={false}
-          />
-          {professorMariOpen && (
-            <span className="absolute -bottom-0.5 left-1/2 h-0.5 w-3 -translate-x-1/2 rounded-full bg-gradient-to-r from-teal-500 to-cyan-500" />
-          )}
-        </button>
-        <span className="mari-chat-title-divider" aria-hidden />
-      </div>
+      <ChatTitleControls
+        className="pl-2.5 pr-0"
+        professorMariOpen={professorMariOpen}
+        onOpenProfessorMari={onOpenProfessorMari}
+        onGoHome={onGoHome}
+      />
       <div
         className="mari-titlebar-content flex h-full min-w-0 flex-1 items-center"
       >
